requirements.txt
tweepy==4.14.0
pandas==2.1.1
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
scikit-learn==1.3.2 